+++ integration/app/src/examples/count/count-sub.state.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { action, debounce, named, payload, Persistence, StateRepository } from '@ngxs-labs/data/decorators';
+import { DataAction, debounce, named, payload, Persistence, StateRepository } from '@ngxs-labs/data/decorators';
 import { NgxsImmutableDataRepository } from '@ngxs-labs/data/repositories';
 import { PersistenceProvider } from '@ngxs-labs/data/typings';
 import { State } from '@ngxs/store';
@@ -17,7 +17,7 @@ const options: PersistenceProvider[] = [{ path: 'count.countSub.val', existingEn
 @Injectable()
 export class CountSubState extends NgxsImmutableDataRepository<CountModel> {
     @debounce()
-    @action()
+    @DataAction()
     public setDebounceSubValue(@payload('value') @named('val') val: number): void {
         this.ctx.setState({ val });
     }