+++ integration/app/src/examples/todo/todo.state.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { action, payload, Persistence, StateRepository } from '@ngxs-labs/data/decorators';
+import { DataAction, payload, Persistence, StateRepository } from '@ngxs-labs/data/decorators';
 import { NgxsImmutableDataRepository } from '@ngxs-labs/data/repositories';
 import { Immutable } from '@ngxs-labs/data/typings';
 import { State } from '@ngxs/store';
@@ -12,14 +12,14 @@ import { State } from '@ngxs/store';
 })
 @Injectable()
 export class TodoState extends NgxsImmutableDataRepository<string[]> {
-    @action()
+    @DataAction()
     public addTodo(@payload('todo') todo: string): void {
         if (todo) {
             this.ctx.setState((state: Immutable<string[]>): Immutable<string[]> => state.concat(todo));
         }
     }
 
-    @action()
+    @DataAction()
     public removeTodo(@payload('idx') idx: number): void {
         this.ctx.setState(
             (state: Immutable<string[]>): Immutable<string[]> =>